requirements.txt
# Web Framework
Flask==2.0.1

# Speech Recognition
openai-whisper==20230314
torch==2.0.1
torchaudio==2.0.1

# Video Processing
ffmpeg-python==0.2.0

# Additional Dependencies
numpy==1.24.3
pydub==0.25.1
requests==2.28.2
Werkzeug==2.2.3
gunicorn==20.1.0